requirements.txt
flask
flask-socketio
eventlet
clipspy
